Correctly synchronize configuration initialization
By putting the null check of the `globalSettings` static variable
outside of the synchronization lock, two threads could both find the
variable null, try to take the lock, and then initialize the variable in
sequence.

In this case, since it's reading a static file to create a `Properties`
object, the duplicate initialization is harmless, but the code is
incorrect.

Put the null check behind the synchronization lock to ensure the
variable is only initialized once.

This error was raised by Spotbugs.

Issue #915 Use static analysis to find bugs