Remove unused, non-functional class
This class was apparently intended to be used with the provider type
controller, but it has no functionality and is never referenced (as
verified with both IntelliJ and `git grep`). Delete it.

Spotbugs report an unwritten field in the class; further analysis
revealed the class was unused entirely.

Issue #915 Use static analysis to find bugs