Add basic security interceptor
Our FHIR API had no authentication layer, meaning anybody with
access to the server could access all data without having an account.
This commit leverages FHIR interceptors (http://hapifhir.io/doc_rest_server_security.html)
which basically requires users to log in using HTTP Basic authentication
in order to use any of the FHIR APIs.

Authorization (e.g. what resources a given user can access) will be
handled at a different interceptor.

It is worth noting that OAuth 2.0 was explored as an authentication
option.  Spring security has support for OAuth, but it was going to
be a heavier lift without much immediate payoff.  Other forms of
authentication will be easy to add; simply create a different
interceptor with that logic and swap it in.

I also spent some time researching the potential to use existing
spring security logic instead of the basic auth interceptor. There
is potential for this tactic, but in the interest of time I suggest
we persue it down the line.

Once a user has been authenticated as a valid user, the user's
permissions need to be checked to ensure they have access to
the resources they are working with.  For now the business logic
for authorization is high level: users have read or write access
to the entire API.

In future there may be more granular rulesets.  FHIR provides the
ability to define more complex rules, and it is possible to
either replace this interceptor or add on top of it as requirements
for authorization change.

Issue #847 Add authn/authz to FHIR API