Change LeieAutomaticScreening matches to an eager set.
Related to Issue #759: Automatically check SSN against the Death Master File

This is required to add future types of automatic screenings.
If each have their own set of matches, we can no longer use
NamedEntity annotations because of overlapping, meaning we
need to eagerly fetch them.  Eagerly fetching multiple inherited
entities can't be done with lists, and has to be done with sets.

This change lays the groundwork for those future changes.