Add dependency on guava
From their documentation[1]:

    Guava is a set of core libraries that includes new collection types
    (such as multimap and multiset), immutable collections, a graph
    library, functional types, an in-memory cache, and APIs/utilities
    for concurrency, I/O, hashing, primitives, reflection, string
    processing, and much more!

A few people on the team have used this library in other projects, and
we've been missing it in this one. Each time we've wanted it, we thought
it was too small a reason to add a dependency, but we think there have
been enough times now that we should just add it so that we'll have it
available the next time we want to use it.

[1] https://github.com/google/guava