Repurpose screening schedules for rescreenings
The existing screening schedule UI and model has no functionality
backing it. Adapt it for the monthly rescreening feature.

Automatic monthly re-screening could mean a bunch of different things;
for this first pass, it means that on a given day of the month, at a
given hour of that day, we will start to do automatic rescreenings. That
functionality is not in this commit, but the UI and database backing is.

Limit the day of the month to be in the range [1, 28], so that
rescreenings still happen in February; limit the hour of the day to be
in the range [0, 24). For the UI, use drop-down menus for input so as to
simplify client-side validation. If users use the dev tools or similar
to submit bad data, throw an exception.

Issue #127 Implement screening schedules or remove UI
Issue #740 Re-screen approved providers monthly