Re-screen approved enrollments monthly
Run our automatic screenings against external data sources on a monthly
basis.

We rely on the container to invoke the `@Schedule`d method hourly, and
then we check if this matches the screening schedule saved in the
database. This is the simplest way to have a user-configurable screening
schedule, but results in some extra work in the form of hourly schedule
look-ups.

Issue #740 Re-screen approved providers monthly