Ignore enrollments w/out risk levels in report
Not all enrollments have risk levels. This is a somewhat erroneous state
that the application can get into; risk levels are supposed to be
assigned shortly after an enrollment is submitted, but if an error
happens and a risk level is not assigned, this report should still work.

Ignore enrollments that do not have a risk level assigned.

Issue #786 Add report based on CMS' provider risk level