Handle missing affiliations in API transformer
Due to the structure of our data model and how we load it, it is
possible for the list of affiliations on a provider profile to be null.
If that is the case, iterating over that list causes a null pointer
exception. Instead, check for null and, if found, return an empty list.

Issue #846 Add new fields to FHIR API
PR #905 Add provider Address and Telecom fields to PSM FHIR API