Update tests for renamed transformer
The commit 022f8292b2dd2f4cfce5e7906b0982f4180094d2 renamed
`EntityToFhir` to `ProviderProfileToFhir`, and updated its type
signature to match, but did not update the tests to pass in objects of
the correct type.

Update the tests to call the transformer with a `ProviderProfile`
containing the previously-expected `Entity` subclass, so that they pass
once more.

Note that this does not add tests for the newly-added address and phone
number functionality; this is just to get the tests passing again.

Issue #846 Add new fields to FHIR API
PR #905 Add provider Address and Telecom fields to PSM FHIR API