Remove checked exception from searchPractice
The problem with checked exceptions is that they affect everything
upstream of them. In this case, the exception was declared but never
thrown; remove the declaration from
ProviderEnrollmentService#searchPractice, and unwrap the
now-unnecessary try-catch blocks from all the upstream call sites that
no longer need to handle a potential PortalServiceException.