Put participation information in one place
Move the section on participation from README.md to CONTRIBUTING.md,
so that all information about how to engage with the project is in one
place.  The proximate motivation for this change is that the CAV ETL
project is about to refer to this documentation, but it's a desirable
change in its own right anyway.

Note that in doing the move, I deliberately dropped an obsolete blurb
about team meeting notes.