Move registration init from base class to derived
Move the registration service initialized in the base system admin
controller to the two derived classes, so that their dependencies are
more explicitly obvious, and so that we will be able to soon remove the
base class. Rely on Spring to inject the service as a constructor
argument, rather than a declared property.