Move standalone test app into test source set
This test application should not be included when building the PSM, but
may yet be useful in verifying that our Drools rules work as expected.
If and when we want more validation of our rules, we should convert this
app into Spock tests.

Move DLSTester out of the main source set and into the test source set.

Co-authored-by: Marvin Wolfthal <maw@weichi.com>