Update liquibase gradle plugin to 2.0
The gradle plugin for liquibase that we were using got updated
to a [2.0
release](https://github.com/liquibase/liquibase-gradle-plugin#july-14-2018)
.  The big benefit of the new version is that liquibase
is no longer packaged as a dependency.  This provides much more flexibility
around the version of liquibase that actually gets used by the project, as
we can pick it ourselves.

As a result we needed to add liquibase as a dependency, and move the database
dependency to the db project.

There are also some changes around which root directory gets cited; resulting
in required changes to the way changesets and seed files are referenced across
the migration-related paths.

Note: we are not using the latest version of liquibase (3.6.x) because
it seems to introduce a change that may make our pending seed data
improvement more verbose than we would like.  I've filed an issue in
liquibase JIRA [for their consideration](https://liquibase.jira.com/browse/CORE-3287)

Issue #980 Improve the way seed data is migrated
Issue #981 Update liquibase gradle plugin to 2.0