Test personal email address displays as required
On both the 'Register New' and the edit existing enrollment
application pages, (for individual providers), > personal info tab.
Display the 'Personal Email Address' input field as required when
the 'Same as Above' input field is checked.

Before this commit, the 'personal email address' input field was
displayed as optional when the 'same as above' input field was
checked.

This commit tests the 'personal email address' input field displays
as required when the 'same as above' input field is checked. This
commit tests the 'Personal Email Address is required when same as
above is checked.' error message displays as expected.

Fixes #842 Show error on personal email address