Display personal email address as required
On both the 'Register New' and the edit existing enrollment
application pages, (for individual providers), > personal info tab.
Display the 'Personal Email Address' input field as required when
the 'Same as Above' input field is checked.

Before this commit, the 'personal email address' input field was
displayed as optional when the 'same as above' input field was
checked.

This commit displays the 'personal email address' input field as
required when the 'same as above' input field is checked.

Fixes #842 Show error on personal email address