Remove service categories table
There were two identical tables: 'categories_of_service' and
'service_categories'.  This was accidental, but both were used in
various portions of the code base.

This commit removes the lesser-used service_categories table, and shifts
all references to the categories_of_service table.

In addition to being more used in the code, most of the dicussion and
documentation of the categories of service feature referenced it in this
form.

Issue #986 Consolidate "service categories" and "categories of service"