Remove the link from accepted_agreemnts to profile.
There was logic around ensuring that previously accepted agreements were
transfered over when making changes to an enrollment.  That has been
removed, so that each time you modify and resubmit an enrollment, you
have to agree to all the agreements anew.

Issue #949: Give providers the ability to "edit and resubmit" applications