Update renewal ticket creation for new provider/enrollment relatonship.
When creating a renewal, it now creates a new enrollment, which is then
set to pending.  Because creating a renewal reuses the same code as
creating the enrollment in the first place, we need to switch on whether
the profile was previously added (has an id of 0 or not) in order to
decide whether to create a new one.  Similarly, selecting a profile
where ticketId was zero no longer made sense and broke the renewal
process.

Issue #949: Give providers the ability to "edit and resubmit"
            applications.