Update search results to use new provider/enrollment link.
In the new system, it no longer makes sense to use the marker IDs of
zero, because these should always be set.  We can also just use the fact
that the two tables are linked explicitly by profile_reference_id.

Issue #949: Give providers the ability to "edit and resubmit" applications