Change Enrollment and UserRequest ticketId to enrollmentId.
The enrollments table already uses enrollment_id, and this brings the
Enrollment class in line with that.  Even though a change from
enrollment to application is coming, this is still a good step in the
right direction.  Because a UserRequest is kind of a hybrid
Provider/Enrollment object for the purpose of listing out enrollments,
it should be changed there too.

Related to Issue #485: Application vs. Enrollment