Rebuild how provider and enrollment interact.
This changes the provider_profile to be merged rather than created when
doing various system things, and to maintain a link with the enrollment
as it moves through the review process.  Before, the process was to blow
away the profile in the database, then create a brand new one, during
certain enrollment updates (like renewal, or draft -> pending).  Instead
of that, we continue to purge and re-add the other parts of the data
model, but update the provider_profile.  This ensures that as new
enrollments are added for this provider_profile, old ones are not
invalidated as it gets updated in the future.

This also removes the duplication of control_no and profile_id,
collapsing the fields into one profile_id which should never be set to 0

This provides the groundwork for separating the rest of the tables to be
either pertinent to enrollments or providers.

The new entities.csv is needed because entities need to link to a
provider explicitly, and the test data linked to the provider with id 0

Issue #949: Give providers the ability to "edit and resubmit" applications