Rebuild how provider and enrollment interact.
This changes the provider_profile to be merged rather than created when
doing various system things, and to maintain a link with the enrollment
as it moves through the review process.

This also removes the duplication of control_no and profile_id,
collapsing the fields into one profile_id which should never be set to 0

This provides the groundwork for separating the rest of the tables to be
either pertinent to enrollments or providers.

Issue #949: Give providers the ability to "edit and resubmit" applications