Migrate data for existing database.
Because the way the links has been specified, we need to migrate how the
enrollments and profiles link to each other.  This can't really be
rolled back, but we include an empty rollback because even though the
state of the database after rollback won't work with the state of the
app after rollback, you can still roll forward again, and if you didn't
have legacy data, the update will be a noop.

Issue #949: Give providers the ability to "edit and resubmit" applications