Fix profile queries.
When updating the way the enrollments and providers interact, the
correct way to see what the active profiles are are to just look for
ones where the owner_id is set to the current user.  Similarly, we need
to allow approvals of updates and renewals, and ensure that those are
getting set after those actions are taken.

Issue #949: Give providers the ability to "edit and resubmit" applications