Remove provider statement date
The last page of the enrollment application process is for the user to
sign a provider statement. Previously, we included the date on that
page, and required the user to type in today's date.

It turns out this is neither necessary nor useful. We don't require it
for legal reasons. It's bad UX, because there's only one acceptable
answer, and the system knows what it is.  Additionally, if an admin
needs to later edit the application, they were sometimes forced to
update the signature date, which is both annoying and incorrect.

Remove the signing date entirely: do not ask for it, do not test that it
is a valid value, do not record it, and do not display it. Instead, we
can rely upon the application submission date.

Issue #162 Provider enrollment: "Provider Statement" UI improvements
Issue #301 When service admin edits pending enrollment, PSM
           inappropriately checks provider statement date
Issue #1021 Remove date field from provider disclosure forms