Remove unused CSS
The `providerPic` CSS class is not used anywhere, and as far as I can
tell, never has been. We deleted the second-to-last reference to it
as part of removing the provider statement signing date (because we
deleted the enclosing class, which was not used anywhere else); delete
the last reference.