Replace old pagination templates with new ones on Functions pages
And fix a problem in the "Displaying A - B of C" where the C was
missing on the service agreements page and the provider types
page. Also fix the drop down 'Show X' menu so it works on the
service agreements page.

Issue #1007: Glitches in pagination widget for service admin tables