Move the dashboard into the static site and remove iframe
This fixes psm-dashboard issue #33 ("Burn-down chart pop-ups don't
work well inside iframe"), and moves the frontend / client-side
files to their new canonical location with the rest of the
projectpsm.org files.

This commit leaves the dashboard page as a static html file, a
hard-coded special case. We can improve on this in follow-up
work, after getting this fix live on the site. Similarly, there
is more to do to remove pym.js in follow-up work.