Remove cms-portal-services:test from CI docs
The cms-portal-services subproject does not have any tests, nor code for
it to test; it is a place to collect all the other code for deployment,
along with a bit of configuration. Invoking the `test` target was always
a no-op.

Pull request #996 updates the project configuration so that the `test`
target is no longer available in the `cms-portal-services` project;
after removing it from our CI configuration (which is not contained in
this repository), also remove it from our documentations.

Issue #629 Use Jenkins for continuous integration
PR #996 Revamp the build process to integrate with eclipse IDE for Hot
        Deploy