Use Spring annotations to find form validators
Now that we've configured Spring to look for annotated controllers
instead of explicitly setting them in XML, we can use annotations to
find other classes Spring is in charge of initializing, too. Apply the
`@Component` annotation to the various form validators in our
application.

Spring describes the `@Component` annotation as "a generic stereotype
for any Spring-managed component"[1].

[1] https://docs.spring.io/spring-framework/docs/current/spring-framework-reference/core.html#beans-stereotype-annotations