Fix suppressed warnings caused by EMPTY_LIST
The `EMPTY_LIST` field of the `Collections` helper class is not
type-safe, and the `emptyList()` method was introduced in Java 1.5 as a
type-safe alternative.

Replace all our uses of `EMPTY_LIST` with `emptyList()`, and remove the
now-unneeded unchecked cast warning suppresssions the type-unsafe field
necessitated.

Issue #915 Use static analysis to find bugs