Make table rows striped with subtle alternating background color
The provider enrollment tables already has this styling; use it
in other tables. Also add a subtle border to make rows easier to
scan. These changes help when links in action columns wrap to two
lines, making it easier to see which row the links belong to.

Use a `linedTable` class to add the lines to target those styles
to only certain tables. At a later point it may make sense to
merge these styles into the `generalTable` style if they make
sense in all `generalTable` cases.

(The `fixedWidthTable` class was causing the right border of that
table to be hidden behind the alternating row backgrounds, so
remove it.)

Issue #692: Action column may wrap to two lines