Add test data to mock out DMF API
The DMF API in the cavetl repo is difficult to set up, mostly because it
requires special access and permissions and so on. For the sake of
demos, and to ease testing the PSM side of the DMF integration, add some
test data and document how to use it.

We want this mock data in the PSM repository and not in the CAVETL
repository for a few reasons:

- this is, like the code that consumes the DMF API, documentation of
  what the PSM expects
- these mocks are useful for testing and demonstrating the PSM
- these mocks are useful for writing integration tests
- the PSM's dependency on CAVETL does not include explicit version
  numbers or commit hashes, and if the DMF API implementation in the
  CAVETL repo changes, the PSM should be able to fall back to these
  mocks until it is updated to consume the new API (and these mocks
  should be updated to match)
- as a side effect of the previous point, it may be useful to be able to
  go back in time to previous versions of the PSM, and having these
  mocks eases that

Thanks to Frank for sharing the special SSNs and names returned by the
Veris Test API.

Issue #759 Automatically check SSN against the Death Master File
Issue #1050 Add more integration tests

Co-authored-by: Frank Duncan <frank@kank.net>