Fix enrollment filter features
Print, PDF export, and filters are now working.

Unfortunately the overall implementation is a little rough.  This is
because the functionality of this form is defined in a common script
that has hard-coded references to a series of enrollment forms.  Ideally
we would refactor this implementation into something more modular and
self-contained.

Issue #1014 Admin UI: Print and Export-To-PDF buttons fail (either showing no applications or showing too many).