Add seed data for inhabited US territories
The United States of America has five inhabited territories, and they
have Medicaid programs, too. Add the territories to the list of "states"
(and I feel like there should be a better word for that table now, but I
haven't been able to find it yet).

Issue #62 Include territories in list of states