Sort lookups, esp list of states and territories
The helper method for finding all the rows of a given lookup type in the
database is used in several places, and wasn't previously sorting the
results, instead returning them in the order the database felt best.
Instead, specify that they should be sorted by name (the `description`
field on each table), primarily so that the list of states and
territories is in the correct order.

Issue #62 Include territories in list of states