Update minimum sequence value for hibernate_sequence.
This allows us to pre load data with primary key values under 2000
without running into conflicts as hibernate adds objects to the
database.

This is a highly irregular change of the first changeset due to the fact
that you will run into PK errors with later data that's inserter for
test purposes if you build a new DB and then start adding enrollments.

Because anyone who has already created the hibernate_sequence shouldn't
have this issue as they are humming happily along, and anyone who is
starting from scratch should have this set, the change was made to an
already existing changeset.

This should usually not be done, but it's the most elegant want to
ensure existing databases aren't affected, but similarly allowing things
to run well for new instances.