Fix positioning of closing "</a>" tags
The immediate reason for this change was to get rid of a spurious
space before a period, in the HTML output: "PSM issue tracker ."
But I fixed all of the "</a>" tags, not just that one, because in
general we don't want to encourage those extra spaces.  Even though
they're not visible right now in most cases, they could unexpectedly
become visible through someone making an innocent edit.