Fix #1105: Access denied error on save and submit
When adding an enrollment, the ownerId was not being set, because that
was a marker flag for which applications have been approved.  However,
with the new refactor, that is used for permissions checks due to the
fact that draft provider_profiles are no longer a thing.  This commit
sets the ownerId for every profile created, but will require looking at
the My Profiles page in the future.

Issue #949: Give providers the ability to "edit and resubmit" applications