Add OwnerId to translation to/from xml.
The provider profile's ownerid wasn't making it through edits of
enrollments because it wasn't being saved to the correlating xml entity,
and so was lost the next time the enrollment was persisted.

Also uncommented ignored tests that were failing due to this.

Issue #949: Give providers the ability to "edit and resubmit" applications